requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
fastapi==0.110.1
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
idna==3.7
imageio==2.34.0
Jinja2==3.1.3
joblib==1.4.0
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.4.18
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
ultralytics==8.1.45
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
wrapt==1.14.1
